% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_choose_docker.R
\name{auto_choose_docker}
\alias{auto_choose_docker}
\title{Automatically choose an appropriate Docker image
to run po4a}
\usage{
auto_choose_docker()
}
\value{
Character vector of length 1
}
\description{
Automatically choose an appropriate Docker image
to run po4a
}
\examples{
auto_choose_docker()
}
