% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_locale.R
\name{create_locale}
\alias{create_locale}
\title{Create the folder and file hierarchy needed to translate a lesson}
\usage{
create_locale(lang)
}
\arguments{
\item{lang}{Name of language. Should be formatted as
\href{https://en.wikipedia.org/wiki/List_of_ISO_639-1_codes}{ISO 639 two-letter language code},
OR the language code followed by an underscore and the
\href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2}{ISO 3166 two-letter country code}
for region-specific translations. Should be
lowercase.}
}
\value{
Nothing; externally, files will be written to \code{locale/{lang}}.
}
\description{
This will create a folder called \code{locale/{lang}}, where \code{{lang}} is provided
via the \code{lang} argument, and populate it with all files needed to translate
into that language; these are initially \strong{copies} of the files in the
original language.
}
\examples{
\dontrun{
# This example assumes the command is being run from the root of a
# carpentries lesson.
create_locale("ja")
}
}
