% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_po.R
\name{create_po}
\alias{create_po}
\title{Create a single PO file for translating a (R)markdown file}
\usage{
create_po(md_in, po, wrap_width = 0, other_args = NULL)
}
\arguments{
\item{md_in}{Character vector of length 1; path to existing MD file.}

\item{po}{Character vector of length 1; path to PO file to write.}

\item{wrap_width}{Integer vector of length 1; number of characters
to wrap lines in PO file. \code{0} (default) indicates no wrapping.}

\item{other_args}{Character vector; other arguments used by
\href{https://mdpo.readthedocs.io/en/master/cli.html#md2po}{md2po}.
Must be formatted like the \code{args} command used by \code{\link[processx:run]{processx::run()}}.}
}
\value{
List with components formatted like the output of \code{\link[processx:run]{processx::run()}}.
Externally, the PO file will be generated at the path specified by \code{po}.
}
\description{
Create a single PO file for translating a (R)markdown file
}
\examples{
\dontrun{
# This example assumes the command is being run from the root of a
# carpentries lesson.
create_po("episodes/01-introduction.Rmd", "po/ja/01-introduction.po")
}
}
