% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_po_for_locale.R
\name{create_po_for_locale}
\alias{create_po_for_locale}
\title{Create all PO files needed for a given locale}
\usage{
create_po_for_locale(lang, container_id = auto_choose_docker())
}
\arguments{
\item{lang}{Name of language. Should be formatted as
\href{https://en.wikipedia.org/wiki/List_of_ISO_639-1_codes}{ISO 639 two-letter language code},
OR the language code followed by an underscore and the
\href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2}{ISO 3166 two-letter country code}
for region-specific translations. Should be
lowercase.}

\item{container_id}{Character vector of length 1;
name of Docker container that includes po4a to run.}
}
\value{
Nothing; externally, one or more PO files will be written to
\code{./po/{lang}}.
}
\description{
This will create a folder called \code{po/{lang}}, where \code{{lang}} is provided via
the \code{lang} argument and populate it with all PO files needed to translate
into that language.
}
\examples{
\dontrun{
# This example assumes the command is being run from the root of a
# carpentries lesson.
create_po_for_locale("ja")
}
}
