% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdpo.R
\name{md2po}
\alias{md2po}
\title{Create a PO file from an MD file}
\usage{
md2po(md_in, po, wrap_width = 0, other_args = NULL)
}
\arguments{
\item{md_in}{Character vector of length 1; path to existing MD file.}

\item{po}{Character vector of length 1; path to PO file to write.}

\item{wrap_width}{Integer vector of length 1; number of characters
to wrap lines in PO file. \code{0} (default) indicates no wrapping.}

\item{other_args}{Character vector; other arguments used by
\href{https://mdpo.readthedocs.io/en/master/cli.html#md2po}{md2po}.
Must be formatted like the \code{args} command used by \code{\link[processx:run]{processx::run()}}.}
}
\value{
List with components formatted like the output of \code{\link[processx:run]{processx::run()}}.
Externally, the PO file will be generated at the path specified by \code{po}.
}
\description{
This is a wrapper around the
\href{https://mdpo.readthedocs.io/en/master/cli.html#md2po}{md2po}
command. It calls \code{md2po} from Docker,
and therefore requires Docker to be installed and running.
}
