% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{print_tag}
\alias{print_tag}
\title{Print the div tag and text}
\usage{
print_tag(block, previous = NULL, parent = NULL)
}
\arguments{
\item{block}{a roxygen2 tag}

\item{previous}{the value of the previous roxygen tag}

\item{parent}{the value of the overall parent tag}
}
\value{
a character vector
}
\description{
Print the div tag and text
}
\keyword{internal}
