% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_trans}
\alias{render_trans}
\title{Render a translated lesson}
\usage{
render_trans(
  lesson_dir = ".",
  l10n_branch = "l10n_main",
  main_branch = "main",
  lang,
  clean = TRUE,
  preview = TRUE,
  include_glob = NULL
)
}
\arguments{
\item{lesson_dir}{Path to the lesson directory. Default: current working
directory.}

\item{l10n_branch}{Name of branch with translated files
(the "translation branch").}

\item{main_branch}{Name of main branch (with files in the source language).}

\item{lang}{Language code.}

\item{clean}{Logical; should any existing local translation branch
be deleted before and after rendering? If \code{TRUE}, a fresh copy of the
translation branch will be checked out from the remote. Default: \code{TRUE}.}

\item{preview}{Logical; should the translated webpage be opened in a
browser window? Default: \code{TRUE}.}

\item{include_glob}{Character; only files with names matching one or more
strings in this character vector will be translated. Use to limit the
files for translation. Default: \code{NULL} (all files with translations
available will be translated).}
}
\value{
Nothing. This is typically called for its side-effect of rendering
a translated webpage. The webpage will be written to the \code{site} folder of
the lesson.
}
\description{
Render a translation of a Carpentries lesson built with
\href{https://github.com/carpentries/sandpaper}{sandpaper}.
This requires that the lesson repo have a translation branch containing
translated files (typically maintained via Crowdin).
}
\details{
The translated files should be located in a \verb{./locale/<LANG_CODE>/} hierarchy
on the translation branch, for example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{.
└── locale
    ├── es
    │   ├── README.md
    │   ├── episodes/
...
}\if{html}{\out{</div>}}

where \code{README.md} and the content of \verb{episodes/} are translated files in the
target language (here, Spanish, as indicated by the language code \code{es}).
}
