% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_trans_from_branch}
\alias{render_trans_from_branch}
\alias{render_trans_from_dir}
\title{Render a translated lesson}
\usage{
render_trans_from_branch(
  lesson_dir = ".",
  l10n_branch = "l10n_main",
  main_branch = "main",
  lang,
  clean = TRUE,
  preview = TRUE
)

render_trans_from_dir(lesson_dir = ".", lang, preview = TRUE)
}
\arguments{
\item{lesson_dir}{Path to the lesson directory. Default: current working
directory.}

\item{l10n_branch}{Name of branch with translated files
(the "translation branch").}

\item{main_branch}{Name of main branch (with files in the source language).}

\item{lang}{Language code.}

\item{clean}{Logical; should any existing local translation branch
be deleted before and after rendering? If \code{TRUE}, a fresh copy of the
translation branch will be checked out from the remote. If \code{FALSE}, the
existing local translation branch will be used. Default: \code{TRUE}.}

\item{preview}{Logical; should the translated webpage be opened in a
browser window? Default: \code{TRUE}.}
}
\value{
Nothing. This is typically called for its side-effect of rendering
a translated webpage. The webpage will be written to the \code{site} folder of
the lesson.
}
\description{
Render a translation of a Carpentries lesson built with
\href{https://github.com/carpentries/sandpaper}{sandpaper}.
This requires that the lesson repo have a directory (or branch) containing
translated files (typically maintained via Crowdin).
}
\details{
The translated files should be located in a \verb{./locale/<LANG_CODE>/}
hierarchy, for example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{.
└── locale
    ├── es
    │   ├── README.md
    │   ├── episodes/
...
}\if{html}{\out{</div>}}

where \code{README.md} and the content of \verb{episodes/} are translated files in the
target language (here, Spanish, as indicated by the language code \code{es}).
The \verb{locale/} directory must exist either on the main branch or the
translation branch.

\code{render_trans_from_branch()} may be used when the \verb{locale/} directory
exists on the translation branch, and requires that the lesson directory
is a git repo. If \code{clean} is \code{TRUE}, the repo must be in a clean git state.

\code{render_trans_from_dir()} may be used when the \verb{locale/} directory exists
in the current lesson directory. It does not require the lesson directory to
be a git repo, nor does it care about its git status if it is.
}
