% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{roxy_tag_parse.roxy_tag_end}
\alias{roxy_tag_parse.roxy_tag_end}
\alias{roxy_tag_parse.roxy_tag_questions}
\alias{roxy_tag_parse.roxy_tag_solution}
\alias{roxy_tag_parse.roxy_tag_callout}
\alias{roxy_tag_parse.roxy_tag_challenge}
\alias{roxy_tag_parse.roxy_tag_checklist}
\alias{roxy_tag_parse.roxy_tag_discussion}
\alias{roxy_tag_parse.roxy_tag_keypoints}
\alias{roxy_tag_parse.roxy_tag_objectives}
\alias{roxy_tag_parse.roxy_tag_prereq}
\alias{roxy_tag_parse.roxy_tag_testimonial}
\title{Roxygen tags for dovetail}
\usage{
\method{roxy_tag_parse}{roxy_tag_end}(x)

\method{roxy_tag_parse}{roxy_tag_questions}(x)

\method{roxy_tag_parse}{roxy_tag_solution}(x)

\method{roxy_tag_parse}{roxy_tag_callout}(x)

\method{roxy_tag_parse}{roxy_tag_challenge}(x)

\method{roxy_tag_parse}{roxy_tag_checklist}(x)

\method{roxy_tag_parse}{roxy_tag_discussion}(x)

\method{roxy_tag_parse}{roxy_tag_keypoints}(x)

\method{roxy_tag_parse}{roxy_tag_objectives}(x)

\method{roxy_tag_parse}{roxy_tag_prereq}(x)

\method{roxy_tag_parse}{roxy_tag_solution}(x)

\method{roxy_tag_parse}{roxy_tag_testimonial}(x)

\method{roxy_tag_parse}{roxy_tag_challenge}(x)

\method{roxy_tag_parse}{roxy_tag_solution}(x)
}
\arguments{
\item{x}{a tag}
}
\value{
somethng
}
\description{
Roxygen tags for dovetail
}
\examples{
soln <- system.file("inst", "extdata", "example-solution-only.R",
  package = "dovetail"
)
roxygen2::parse_file(soln, env = NULL)
}
