% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{roxy_tag_parse.roxy_tag_challenge}
\alias{roxy_tag_parse.roxy_tag_challenge}
\title{Roxy challenge}
\usage{
\method{roxy_tag_parse}{roxy_tag_challenge}(x)
}
\arguments{
\item{x}{a tag}
}
\value{
somethng
}
\description{
Roxy challenge
}
\examples{
soln <- system.file("inst", "extdata", "example-number-echo.R",
  package = "roxyblox"
)
roxygen2::parse_file(soln, env = NULL)
}
