% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_md_for_locale.R
\name{translate_md_for_locale}
\alias{translate_md_for_locale}
\title{Translate a set of (R)Markdown files using PO files}
\usage{
translate_md_for_locale(lang)
}
\arguments{
\item{lang}{Name of language. Should be formatted as
\href{https://en.wikipedia.org/wiki/List_of_ISO_639-1_codes}{ISO 639 two-letter language code},
OR the language code followed by an underscore and the
\href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2}{ISO 3166 two-letter country code}
for region-specific translations. Should be
lowercase.}
}
\value{
Nothing; externally, translated (R)Markdown files will be written
to \code{./locale/{lang}}.
}
\description{
This calls \code{md2po} from Docker,
and therefore requires Docker to be installed and running.
}
\details{
It also requires that a folder to write the translated (R)Markdown files
exists at \code{./locale/{lang}}, where \code{{lang}} matches that provided
by the \code{lang} argument.

The easiest way to set this up is to use \code{\link[=create_locale]{create_locale()}} first.

It also requires that PO files are located in a folder like this:

\if{html}{\out{<div class="sourceCode">}}\preformatted{|--- po
     \\---ja/
         |---01-introduction.po
         |---index.po
         ...
}\if{html}{\out{</div>}}

Where \code{01-introduction.po} and \code{index.po} are PO files corresponding
to \code{01-introduction.md} and \code{index.md} in the original language.
The original md files must be located as specified by
\href{https://carpentries.github.io/sandpaper/}{sandpaper},
and have unique file names.
Here, \code{ja} indicates that these PO files are for translating to Japanese.

The easiest way to generate PO files meeting these requirements is to use
\code{\link[=create_po_for_locale]{create_po_for_locale()}} first.
}
\examples{
\dontrun{
# This example assumes the command is being run from the root of a
# carpentries lesson and that all PO files and the `locale/{lang}`
# folder has been set up.
translate_md_for_locale("ja")
}
}
