% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crowdin-yml.R
\name{use_crowdin_yml}
\alias{use_crowdin_yml}
\title{Write crowdin.yml}
\usage{
use_crowdin_yml(lesson_dir = ".", overwrite = NULL)
}
\arguments{
\item{lesson_dir}{Path to the lesson directory. Default: current working
directory.}

\item{overwrite}{Logical; if a \code{crowdin.yml} file already exists in
\code{lesson_dir}, should it be overwritten?}
}
\value{
The path to the newly written \code{crowdin.yml} file (invisibly).
This is typically called for its side-effect of writing
\code{crowdin.yml} to the lesson directory.
}
\description{
The \code{crowdin.yml} file should be placed in the root directory of the lesson.
}
\details{
\code{crowdin.yml} is used by Crowdin for specifying how the translated lesson
files should be organized (amongst other Crowdin settings). \code{dovetail}
expects a certain file hierarchy for the translated lesson files, and the
\code{crowdin.yml} file generated by this function will make sure it is set up
correctly.
}
