% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_po.R
\name{create_po}
\alias{create_po}
\title{Create a single PO file for translating a (R)markdown file}
\usage{
create_po(md_in, po, other_args = NULL)
}
\arguments{
\item{md_in}{Character vector of length 1; path to existing MD file.}

\item{po}{Character vector of length 1; path to PO file to write.}

\item{other_args}{Character vector; other arguments used by
\href{https://po4a.org/man/man1/po4a-updatepo.1.php}{po4a-updatepo}.
Must be formatted like the \code{args} command used by \code{\link[processx:run]{processx::run()}}.}
}
\value{
Path to the PO file.
Externally, the PO file will be generated at the path specified by \code{po}.
}
\description{
Any folders that don't yet exist for the output path specified will be
automatically created.
}
\examples{
# Write a PO file based on the example MD file in this package
# to a temporary file, check the first few lines, then delete it
temp_po <- tempfile(fileext = "po")
create_po(
  md_in = system.file("extdata", "test.Rmd", package = "dovetail"),
  po = temp_po,
  other_args = "--nolocation"
)
head(readLines(temp_po))
unlink(temp_po)

}
