% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdpo.R
\name{md2po}
\alias{md2po}
\title{Create a PO file from an MD file}
\usage{
md2po(md_in, po, other_args = NULL)
}
\arguments{
\item{md_in}{Character vector of length 1; path to existing MD file.}

\item{po}{Character vector of length 1; path to PO file to write.}

\item{other_args}{Character vector; other arguments used by
\href{https://po4a.org/man/man1/po4a-updatepo.1.php}{po4a-updatepo}.
Must be formatted like the \code{args} command used by \code{\link[processx:run]{processx::run()}}.}
}
\value{
Character vector of length 1; the path to the PO file specified by
\code{po}. Externally, the PO file will be generated at that path.
}
\description{
This is a wrapper around the
\href{https://po4a.org/man/man1/po4a-updatepo.1.php}{po4a-updatepo}
command. It calls \code{po4a} from Docker,
and therefore requires Docker to be installed and running.
}
\examples{
# Write a PO file based on the example MD file in this package
# to a temporary file, check the first few lines, then delete it
temp_po <- tempfile(fileext = "po")
md2po(
  md_in = system.file("extdata", "test.Rmd", package = "dovetail"),
  po = temp_po
)
tail(readLines(temp_po))
unlink(temp_po)

}
