% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_block}
\alias{parse_block}
\title{Parse A roxygen-formatted block to into a block to be rendered as markdown}
\usage{
parse_block(txt)
}
\arguments{
\item{txt}{a character vector of length 1 that represents a block of text}
}
\value{
the modified block with div tags assigned in the appropriate places
}
\description{
Parse A roxygen-formatted block to into a block to be rendered as markdown
}
\examples{
# Simple example showing that the block can be parsed with knitr in its own
# environment
e <- new.env()
assign("h", list(ello = "hello", ere = "there"), env = e)
txt <- parse_block("
#' @challenge Hello Challenge
#' Say hello
#'
#' @solution olleH Solution
#'
#' ```{r}
h$ello  # variables in the current environment are evaluated
h$ere
getwd() # working directory is the current directory
#' ```
")
tmp <- tempfile(fileext = ".md")
knitr::knit(output = tmp, text = txt, env = e)
file.edit(tmp)

# Example of a more typical block
f <- system.file("extdata", "example-number-echo.R", package = "roxyblox")
txt <- paste(readLines(f, encoding = "UTF-8"), collapse = "\n")
cat(txt)
ptxt <- parse_block(txt)
tmp <- tempfile(fileext = ".md")
knitr::knit(output = tmp, text = ptxt, encoding = "UTF-8", envir = parent.frame())
file.edit(tmp)

# Example of a block with multiple solutions
f <- system.file("extdata", "example-multi-solution.txt", package = "roxyblox")
txt <- paste(readLines(f, encoding = "UTF-8"), collapse = "\n")
cat(txt)
ptxt <- parse_block(txt)
tmp <- tempfile(fileext = ".md")
knitr::knit(output = tmp, text = ptxt, encoding = "UTF-8", envir = parent.frame())
file.edit(tmp)
}
