% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{roxy_tag_parse.roxy_tag_solution}
\alias{roxy_tag_parse.roxy_tag_solution}
\title{Roxy solution}
\usage{
\method{roxy_tag_parse}{roxy_tag_solution}(x)
}
\arguments{
\item{x}{a tag}
}
\value{
somethng
}
\description{
Roxy solution
}
\examples{
soln <- system.file("inst", "extdata", "example-solution-only.R",
  package = "roxyblox"
)
roxygen2::parse_file(soln, env = NULL)
}
