% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_md.R
\name{translate_md}
\alias{translate_md}
\title{Translate a single (R)Markdown file using a PO file}
\usage{
translate_md(
  md_in,
  po,
  md_out,
  wrap_width = 0,
  other_args = NULL,
  quiet = FALSE
)
}
\arguments{
\item{md_in}{Character vector of length 1; path to existing MD file.}

\item{po}{Character vector of length 1; path to PO file to write.}

\item{md_out}{Character vector of length 1; path to write translated MD file.}

\item{wrap_width}{Integer vector of length 1; number of characters
to wrap lines in PO file. \code{0} (default) indicates no wrapping.}

\item{other_args}{Character vector; other arguments used by
\href{https://mdpo.readthedocs.io/en/master/cli.html#md2po}{md2po}.
Must be formatted like the \code{args} command used by \code{\link[processx:run]{processx::run()}}.}

\item{quiet}{Logical; should messages from this function be suppressed?}
}
\value{
Nothing; externally, the markdown file will be translated to
\code{md_out}.
}
\description{
This calls \code{md2po} from Docker,
and therefore requires Docker to be installed and running.
}
\examples{
\dontrun{
# This example assumes the command is being run from the root of a
# carpentries lesson.
  translate_md(
    md_in = "episodes/01-introduction.Rmd",
    po = "po/01-introduction.ja.po",
    md_out = "locale/ja/episodes/01-introduction.Rmd"
  )
}
}
